+++ src/VBox/Main/src-client/ConsoleImpl2.cpp
@@ -1,4 +1,4 @@
-/* $Id: ConsoleImpl2.cpp 45890 2013-05-03 09:57:44Z noreply@oracle.com $ */
+/* $Id: ConsoleImpl2.cpp 45927 2013-05-07 08:05:43Z noreply@oracle.com $ */
 /** @file
  * VBox Console COM Class implementation - VM Configuration Bits.
  *
@@ -4943,13 +4943,13 @@ static void configSetProperty(VMMDev * const pVMMDev,
     AssertPtrReturnVoid(pszFlags);
     parms[0].type = VBOX_HGCM_SVC_PARM_PTR;
     parms[0].u.pointer.addr = (void *)pszName;
-    parms[0].u.pointer.size = strlen(pszName) + 1;
+    parms[0].u.pointer.size = (uint32_t)strlen(pszName) + 1;
     parms[1].type = VBOX_HGCM_SVC_PARM_PTR;
     parms[1].u.pointer.addr = (void *)pszValue;
-    parms[1].u.pointer.size = strlen(pszValue) + 1;
+    parms[1].u.pointer.size = (uint32_t)strlen(pszValue) + 1;
     parms[2].type = VBOX_HGCM_SVC_PARM_PTR;
     parms[2].u.pointer.addr = (void *)pszFlags;
-    parms[2].u.pointer.size = strlen(pszFlags) + 1;
+    parms[2].u.pointer.size = (uint32_t)strlen(pszFlags) + 1;
     pVMMDev->hgcmHostCall("VBoxGuestPropSvc", guestProp::SET_PROP_HOST, 3,
                           &parms[0]);
 }
